refactor resolvers to prepare for external resolver
