+++ .go-version
@@ -1,1 +1,1 @@
-1.13.7
+1.14.6
