+++ go.mod
@@ -10,7 +10,5 @@ require (
 	github.com/pkg/errors v0.9.1
 	github.com/smartystreets/goconvey v1.6.4
 	github.com/spf13/cobra v0.0.5
-	github.com/sqs/goreturns v0.0.0-20181028201513-538ac6014518 // indirect
-	golang.org/x/tools v0.0.0-20200207001614-6fdc5776f4bb // indirect
 	gopkg.in/yaml.v2 v2.2.2
 )
